requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
google-auth==2.15.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
